pyproject.toml
[project]
name = "ufanet-practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asgiref==3.8.1",
    "django==5.2.1",
    "pillow==11.2.1",
    "psycopg>=3.2.9",
    "psycopg2==2.9.10",
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "pytest-django>=4.11.1",
    "sqlparse==0.5.3",
    "typing-extensions==4.13.2",
    "tzdata==2025.2",
]
